refactor(config): F1 — app_config_* fan-out (P8)
F1 slice shipped. jackin-config's four app_config_*.rs siblings move
from crates/jackin-config/src/ to crates/jackin-config/src/app_config/.
app_config.rs stays as the coordinator; the four impl-bearing files
become app_config/{mounts,persist,roles,workspaces}.rs; the four
tests files become app_config/{mounts,persist,roles,workspaces}/tests.rs.

Touches (13 files modified/created/deleted, 1 unchanged):
  - created crates/jackin-config/src/app_config/mounts.rs (from
    app_config_mounts.rs, with use crate::app_config::AppConfig ->
    use super::AppConfig)
  - created crates/jackin-config/src/app_config/mounts/tests.rs (verbatim)
  - created crates/jackin-config/src/app_config/persist.rs (with two
    use edits: use crate::app_config::AppConfig -> use super::AppConfig
    and crate::app_config_roles::BUILTIN_ROLES -> super::roles::BUILTIN_ROLES)
  - created crates/jackin-config/src/app_config/persist/tests.rs (verbatim)
  - created crates/jackin-config/src/app_config/roles.rs (use edit)
  - created crates/jackin-config/src/app_config/roles/tests.rs (verbatim)
  - created crates/jackin-config/src/app_config/workspaces.rs (use edit)
  - created crates/jackin-config/src/app_config/workspaces/tests.rs (verbatim)
  - modified crates/jackin-config/src/app_config.rs: doc-comment update
    + four new pub mod declarations
  - modified crates/jackin-config/src/lib.rs: removed four pub mod
    app_config_* lines + updated three pub use paths to use
    app_config::module:: re-exports
  - modified crates/jackin-config/src/editor.rs: repoint
    app_config_persist -> app_config::persist
  - deleted crates/jackin-config/src/app_config_*.rs and their tests
    files (git rm -f) + rmdir of the now-empty subdirectories

Decisions:
  - app_config.rs stays as the only file in crates/jackin-config/src/
    at the module-root level; the four children live under app_config/.
    This matches the playbook spec's 'rename the four app_config_*.rs
    top-level siblings into coordinator + sibling layout' goal exactly.
  - The persist.rs second edit (super::roles::BUILTIN_ROLES instead of
    crate::app_config_roles::BUILTIN_ROLES) was needed because the
    reference crossed the new module boundary; the spec calls this
    out explicitly.
  - The lib.rs re-exports (app_config_mounts -> app_config::mounts,
    etc.) preserve the public surface; no call site change is needed
    outside this file.
  - No file-size-budget or test-layout-allowlist entry is needed:
    none of the moved files are over the 2000L cap, and the new
    app_config/* subdirs use sibling tests.rs (no mod tests{} blocks,
    no inline tests).

Verified:
  - cargo fmt --check clean
  - cargo clippy -p jackin-config --all-targets --all-features
    --locked -- -D warnings: exit 0, no warnings
  - cargo nextest run -p jackin-config: 259/259 pass
  - cargo xtask lint: file-size + test-layout OK; arch section is
    informational only (the three remaining edges are unchanged)
  - behavioral specs runtime-launch and op-picker pass UNMODIFIED

Note on the full-workspace clippy: the spec's 'criterion_main!
(benches)' step + the C1-jackin-host parked state in this stack
cause the full-workspace clippy to fail (jackin-host's deps
reference a non-existent 'jackin_runtime::naming' etc.). The
jackin-config crate is the touched crate for F1 and it compiles
green in isolation; jackin-host is a separate (C1) branch that
the operator will resolve when they merge C1.

Refs: plans/codebase-health-playbook.md#f1--app_config_-fan-out--app_config-coordinator--siblings-p8
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
