+++ docs/content/docs/roadmap/codebase-health-enforcement.mdx
@@ -255,7 +255,7 @@ Used by every C/E slice. Pure move; contents are not rewritten.
 
 ### Phase F — naming normalization (P8)
 
-- [ ] **F1 — `app_config_*` → `app_config/` coordinator + siblings** (`app_config.rs` + `app_config/{mounts,persist,roles,workspaces}.rs`).
+- [x] **F1 — `app_config_*` → `app_config/` coordinator + siblings** (`app_config.rs` + `app_config/{mounts,persist,roles,workspaces}.rs`).
 - [ ] **F2 — TUI stem normalization** — retire `state.rs`/`render.rs` synonyms for the canonical TEA stems; folds into Phase G as each stack migrates.
 
 ### Phase G — unify the TUI runtime (W6; largest)
